rename context toggles to context selectors, because they are not a toggle, but a selector
