+++ src/contexts/teamContext/team-selector.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import {
   Select,
   SelectContent,
@@ -14,13 +14,18 @@ import { TeamContext } from "./team-provider";
 import { api } from "@/utils/api";
 
 export function TeamSelector() {
+  const [isClient, setIsClient] = useState(false);
   const { teamId, setTeamId } = useContext(TeamContext);
   const { data, isLoading } = api.team.findAll.useQuery();
+
+  useEffect(() => {
+    setIsClient(true);
+  }, []);
 
   return (
     <Select defaultValue={teamId} onValueChange={(v: string) => setTeamId(v)}>
       <SelectTrigger className="w-[180px]">
-        <SelectValue placeholder="" />
+        <SelectValue placeholder={isClient ? "Pick team" : ""} />
       </SelectTrigger>
       <SelectContent>
         <SelectGroup>
